Trac #25351: update @sage_wraps test cases
One of the @sage_wraps tests retrieved the source code for a method
modified by this ticket, and just needed to be updated.

The other failing test case appears to have been failing already;
the filename returned should be the filename of the function,
not the filename of the decorator.  I labeled it "known bug".