Trac #27606: an initial crack at aligning series expansions in tables
This code works by adding a 'table' option to the _latex_ method, which
causes the output to have alignment characters embedded.  I think it
would be better for the output to be a list of LaTeX strings when
'table' (or better, 'tabular') is specified, and let the table()
function insert the alignment chars.